Days in Year

Given a single numeric value representing a year, it generates:
1) a vector of dates with all the days within the year, OR
2) the amount of days in the specified year

diy(year, out.type = "seq")

Arguments

year

numeric, the year for which the sequence of days will be generated

out.type

Character indicating the type of result that is given by this function. Valid values are:
-) seq : a vectorial sequence with all the days within the given year
-) nmbr: the number of days in the vectorial sequence with all the days within the given year

Author

Mauricio Zambrano-Bigiarini, mzb.devel@gmail

See also

Examples

## Sequence of daily dates for the year 1961
diy(1961)

## Computing the number of days between in 1961
diy(1961, out.type = "nmbr")